Chat Application
I have spent most of today trying to create this chat application. After using chat GPT to give me some good project ideas i felt this would be one i could  tackle with my current skill set.

It was obvious i would need persistent messages and so a database is needed thankfully my exposure to firebase made this a little easier and with a little poking around i was also able to use firebase not only for a firestore database but also my user login and authentication.

I have set up the authentication using google and so a user simply logs in via google and they can chat away. Doing it this way also allows me to attach the user ID to each message allowing for the comparison of the uid on the message to uid of the currently logged in user, to see if a message is sent or recieved.

I have added some basic styling for now but theres no getting around the fact it looks terrible and i will change that asap. There are some other additions i wish to try out for this project also as i am realising the pure power of firebase once again i wish to play around with some other features it offers me and i will begin on that tomorrow. I cannot wait to get this project full coded up.