Clean up and abstraction.
Today i have been cleaning up the code base a little and optimizing performance.

Firstly i realized that the entire chat room was being re-rendered on every keystroke of the input element, this was because i had placed the component directly as part of the chat room. I abstracted it to its own file which now means the chat room only re-renders when necessary and in this case memoization was not necessary.

I also improved the styling slightly. and removed some of the react-app clutter given with create-react-app.

I will be trying to add features for a friends list as well as having seperate chats with each. I may do that with this project i may make a new one i am undecided at the moment, as this was just to prove to myself i could get this functionality working on my own.